Changes: - move is_required from baseCompositeField to implementers - return empty values instead of raising NotImplementedError (reduces   boilerplate code) - cleanup foreign keys, move logic for single foreign keys to foreign   keys - allow selecting fields in manytomany keys - allow passing an abstract model to through as a template
