+++ edgy/core/db/fields/base.py
@@ -5,7 +5,6 @@ from typing import (
     Any,
     ClassVar,
     Dict,
-    FrozenSet,
     Optional,
     Pattern,
     Sequence,
@@ -16,7 +15,7 @@ from typing import (
 
 from pydantic import BaseModel
 from pydantic.fields import FieldInfo
-from sqlalchemy import Column, Constraint, ForeignKeyConstraint
+from sqlalchemy import Column, Constraint
 
 from edgy.core.connection.registry import Registry
 from edgy.types import Undefined
@@ -25,17 +24,6 @@ if TYPE_CHECKING:
     from edgy import Model, ReflectModel
 
 edgy_setattr = object.__setattr__
-
-FK_CHAR_LIMIT = 63
-
-
-def _removeprefix(text: str, prefix: str) -> str:
-    # TODO: replace with removeprefix when python3.9 is minimum
-    if text.startswith(prefix):
-        return text[len(prefix) :]
-    else:
-        return text
-
 
 class BaseField(FieldInfo):
     """
@@ -140,13 +128,13 @@ class BaseField(FieldInfo):
         """
         Returns the columns of the field being declared.
         """
-        raise NotImplementedError()
+        return []
 
     def clean(self, field_name: str, value: Any) -> Dict[str, Any]:
         """
         Validates a value and transform it into columns which can be used for querying and saving
         """
-        raise NotImplementedError()
+        return {}
 
     def to_model(self, field_name: str, value: Any, phase: str = "") -> Dict[str, Any]:
         """
@@ -185,6 +173,7 @@ class BaseField(FieldInfo):
         # for multidefaults overwrite in subclasses get_default_values to
         # parse default values differently
         # NOTE: multi value fields should always check here if defaults were already applied
+        # NOTE: when build meta fields without columns this should be empty
         if field_name in cleaned_data:
             return {}
         return {field_name: self.get_default_value()}
@@ -260,12 +249,6 @@ class BaseCompositeField(BaseField):
                     cleaned_data_result[sub_field_name_new] = default_value
         return cleaned_data_result
 
-    def get_columns(self, name: str) -> Sequence["Column"]:
-        return []
-
-    def is_required(self) -> bool:
-        return False
-
 
 
 class PKField(BaseCompositeField):
@@ -327,22 +310,18 @@ class PKField(BaseCompositeField):
                 return True
         return False
 
+    def is_required(self) -> bool:
+        return False
+
+
 class BaseForeignKey(BaseField):
     def __init__(
         self,
         *,
-        on_update: str,
-        on_delete: str,
         related_name: str = "",
-        related_fields: Sequence[str] = (),
-        through: Any = None,
         **kwargs: Any,
     ) -> None:
         self.related_name = related_name
-        self.related_fields = related_fields
-        self.through = through
-        self.on_update = on_update
-        self.on_delete = on_delete
         super().__init__(**kwargs)
 
     @property
@@ -368,143 +347,11 @@ class BaseForeignKey(BaseField):
         except AttributeError:
             pass
 
-    @cached_property
-    def related_columns(self) -> Dict[str, Optional[Column]]:
-        target = self.target
-        columns: Dict[str, Optional[Column]] = {}
-        if self.related_fields:
-            for field_name in self.related_fields:
-                if field_name in target.meta.fields_mapping:
-                    for column in target.meta.fields_mapping[field_name].get_columns(field_name):
-                        columns[column.key] = column
-                else:
-                    columns[field_name] = None
-        else:
-            # try to use explicit primary keys
-            if target.pknames:
-                for pkname in target.pknames:
-                    for column in target.meta.fields_mapping[pkname].get_columns(pkname):
-                        columns[column.key] = column
-            elif target.pkcolumns:
-                columns.update({col: None for col in target.pkcolumns})
-        return columns
-
     def expand_relationship(self, value: Any) -> Any:
-        target = self.target
-
-        if isinstance(value, (target, target.proxy_model)):
-            return value
-        return target.proxy_model(pk=value)
-
-    def clean(self, name: str, value: Any) -> Dict[str, Any]:
-        target = self.target
-        retdict: Dict[str, Any] = {}
-        if value is None:
-            for column_name in self.get_column_names(name):
-                retdict[column_name] = None
-        elif isinstance(value, dict):
-            for pkname in target.pknames:
-                if pkname in value:
-                    retdict.update(target.meta.fields_mapping[pkname].clean(self.get_fk_field_name(name, pkname), value[pkname]))
-        elif isinstance(value, BaseModel):
-            for pkname in target.pknames:
-                if hasattr(value, pkname):
-                    retdict.update(
-                        target.meta.fields_mapping[pkname].clean(self.get_fk_field_name(name, pkname), getattr(value, pkname))
-                    )
-        elif len(target.pknames) == 1:
-            retdict.update(
-                target.meta.fields_mapping[target.pknames[0]].clean(self.get_fk_field_name(name, target.pknames[0]), value)
-            )
-        else:
-            raise ValueError(f"cannot handle: {value} of type {type(value)}")
-        return retdict
-
-    def modify_input(self, name: str, kwargs: Dict[str, Any]) -> None:
-        if len(self.target.pknames) == 1:
-            return
-        to_add = {}
-        # for idempotency
-        for column_name in self.get_column_names(name):
-            if column_name in kwargs:
-                to_add[self.from_fk_field_name(name, column_name)] = kwargs.pop(column_name)
-        # empty
-        if not to_add:
-            return
-        if name in kwargs:
-            raise ValueError("Cannot specify a foreign key column and the foreign key itself")
-        if len(self.target.pknames) != len(to_add):
-            raise ValueError("Cannot update the foreign key partially")
-        kwargs[name] = to_add
+        return value
 
     def to_model(self, field_name: str, value: Any, phase: str = "") -> Dict[str, Any]:
         """
         Like clean just for the internal input transformation. Validation happens later.
         """
         return {field_name: self.expand_relationship(value)}
-
-    def get_fk_name(self, name: str) -> str:
-        """
-        Builds the fk name for the engine.
-
-        Engines have a limitation of the foreign key being bigger than 63
-        characters.
-
-        if that happens, we need to assure it is small.
-        """
-        fk_name = f"fk_{self.owner.meta.tablename}_{self.target.meta.tablename}_{name}"
-        if not len(fk_name) > FK_CHAR_LIMIT:
-            return fk_name
-        return fk_name[:FK_CHAR_LIMIT]
-
-    def get_fk_field_name(self, name: str, fieldname: str) -> str:
-        target = self.target
-        if len(target.pknames) == 1:
-            return name
-        return f"{name}_{fieldname}"
-
-    def from_fk_field_name(self, name: str, fieldname: str) -> str:
-        target = self.target
-        if len(target.pknames) == 1:
-            return target.pknames[0]  # type: ignore
-        return _removeprefix(fieldname, f"{name}_")
-
-    def get_column_names(self, name: str) -> FrozenSet[str]:
-        if not hasattr(self, "_column_names") or name != self.name:
-            column_names = set()
-            for column in self.owner.meta.field_to_columns[name]:
-                column_names.add(column.name)
-            if name != self.name:
-                return frozenset(column_names)
-            self._column_names = frozenset(column_names)
-        return self._column_names
-
-    def get_columns(self, name: str) -> Sequence[Column]:
-        target = self.target
-        columns = []
-        for column_name, pkcolumn in self.related_columns.items():
-            if pkcolumn is None:
-                pkcolumn = target.table.columns[column_name]
-            fkcolumn_name = self.get_fk_field_name(name, column_name)
-            fkcolumn = Column(
-                fkcolumn_name,
-                pkcolumn.type,
-                primary_key=self.primary_key,
-                autoincrement=False,
-                nullable=pkcolumn.nullable or self.null,
-                unique=pkcolumn.unique,
-            )
-            columns.append(fkcolumn)
-        return columns
-
-    def get_global_constraints(self, name: str, columns: Sequence[Column]) -> Sequence[Constraint]:
-        target = self.target
-        return [
-            ForeignKeyConstraint(
-                columns,
-                [f"{target.meta.tablename}.{self.from_fk_field_name(name, column.key)}" for column in columns],
-                ondelete=self.on_delete,
-                onupdate=self.on_update,
-                name=self.get_fk_name(name),
-            ),
-        ]
